fix: 시트 동기화 수 응답 오버플로우 감지 추가

+++ src/main/java/gg/agit/konect/domain/club/service/ClubMemberSheetService.java
@@ -97,6 +97,6 @@ public class ClubMemberSheetService {
         long preMemberCount = clubPreMemberRepository.countByClubId(clubId);
         sheetSyncExecutor.executeWithSort(clubId, sortKey, ascending);
 
-        return ClubMemberSheetSyncResponse.of((int)(memberCount + preMemberCount), spreadsheetId);
+        return ClubMemberSheetSyncResponse.of(Math.toIntExact(memberCount + preMemberCount), spreadsheetId);
     }
 }
